+++ backend/api/build.gradle
@@ -2,6 +2,7 @@ buildscript {
     ext {
         springBootVersion = '2.3.3.RELEASE'
         querydslPluginVersion = '1.0.10'
+        mapstructVersion = "1.3.0.Final"
     }
     repositories {
         mavenCentral()
@@ -66,6 +67,12 @@ dependencies {
      */
     compile("com.querydsl:querydsl-jpa")
     compile("com.querydsl:querydsl-apt")
+
+    /**
+     * MapStruct
+     */
+    implementation "org.mapstruct:mapstruct:${mapstructVersion}"
+    annotationProcessor "org.mapstruct:mapstruct-processor:${mapstructVersion}"
 
     /**
      * Lombok
