+++ backend/api/src/main/java/me/travelplan/service/user/UserService.java
@@ -10,17 +10,11 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
     private final UserRepository userRepository;
 
-    public User create(String email, String password, String name) {
-        if (userRepository.findUserByEmail(email).isPresent()) {
+    public User create(User user) {
+        if (userRepository.findUserByEmail(user.getEmail()).isPresent()) {
             throw new RuntimeException("중복된 이메일입니다");
         }
 
-
-        return userRepository.save(User.builder()
-                .email(email)
-                .password(passwordEncoder.encode(password))
-                .name(name)
-                .build()
-        );
+        return userRepository.save(user);
     }
 }
