+++ backend/api/src/main/java/me/travelplan/web/AuthRequest.java
@@ -1,0 +1,18 @@
+package me.travelplan.web;
+
+import lombok.Getter;
+
+public class AuthRequest {
+    @Getter
+    public static class Register {
+        String email;
+        String password;
+        String name;
+    }
+
+    @Getter
+    public static class Login {
+        String email;
+        String password;
+    }
+}
